requirements.txt
pandas==2.2.2
Requests==2.32.3
streamlit==1.36.0
streamlit-option-menu==0.4.0
openai
tensorflow==2.17.1
streamlit
pillow
numpy
scikit-learn
pandas
